refactor(intel): extract architecture-agnostic RecursiveDisassembler engine
Decouple the recursive CFG-recovery engine from x86 so the same method can drive
other native ISAs (ARM/MIPS, issue #2). The engine now lives in
common/RecursiveDisassembler.py and delegates every architecture-specific
decision to an injected ArchBackend (common/arch/ArchBackend.py). The x86
specifics -- capstone setup, the call/jmp/cond-jmp/loop/ret classification and
the push-ret / syscall-exit / alignment-after-call idioms -- move verbatim into
intel/X86Backend.py. IntelDisassembler collapses to a thin back-compat wrapper
injecting X86Backend.

FunctionCandidateManager, FunctionAnalysisState, definitions and the analyzers
are unchanged; the engine receives them through backend factory methods.

No behavior change: full suite passes identically (111 passed, 79 subtests),
including testIntegration's exact report sha256 on the Asprox and Cutwail
samples (byte-for-byte identical output). Timing on asprox is within <1% of
master across min/mean/median.
